able to delete draws that were uploaded

+++ routes/invoices.js
@@ -229,9 +229,14 @@ router.post('/update', checkAuth, async (req, res) => {
  * Purpose: Delete an invoice
  */
 router.post('/delete/:projectId/:draw', checkAuth, async (req, res) => {
-  var fileName = /[^/]*$/.exec(req.body.invoicePath)[0]
-  let result = await deleteFile(fileName)
-  console.log(result)
+  console.log(req.body)
+  console.log(req.body.invoicePath)
+  if (req.body.invoicePath.length > 0) {
+    var fileName = /[^/]*$/.exec(req.body.invoicePath)[0]
+    let result = await deleteFile(fileName)
+    console.log(result)
+  }
+
 
   Project.findOneAndUpdate(
     {
@@ -264,6 +269,7 @@ router.post('/upload/:projectId/:draw', (req, res) => {
   let body = req.body;
   const currentDateTime = new Date().toISOString();
   let invoices = body.map(v => ({...v, _id: mongoose.Types.ObjectId(), dateEntered: currentDateTime}))
+  console.log(JSON.stringify(invoices))
 
   Project.findOneAndUpdate(
     {
